Issue #12, #14: Add Company and CompanyCategory entities
I have added Company and CompanyCategory entities in the same 'Company' domain as they are related and CompanyCategory is not a core business domain. I also set spring.jpa.hibernate.ddl-auto to update (This is only recommended in dev environment. DO NOT USE update or create or create-drop in production. This should be set to validate or none in staging or production